refactor(htmlcss): one shared parse-and-cascade front-end
The five-step csscascade sequence (parse → CascadeDriver::new →
bootstrap_dom → flush → style_document) existed verbatim in both
collect_styled_tree (renderer) and import::html::from_html_str
(importer). It now lives once, in htmlcss/frontend.rs
(parse_and_style), returning the styled HtmlDocument handle.

The layout.grid.enabled pref stays renderer-side in
collect_styled_tree — deliberately NOT part of the shared front-end —
so the importer's Stylo behavior is bit-for-bit what it was.

Gates: html_import_snapshot — zero golden diffs across all 139 L0
fixtures; full suite green (32 suites); L0.exact htmlcss render
byte-identical to the M0 baseline (gridaco/nothing#27).

Part of gridaco/nothing#30, under the seam program gridaco/nothing#27.
